+++ lib/src/command_manager.dart
